Version 1.00, BuildID 1.38 * Set arm button to pull up * Added del file command * Prevent bricking when flashing bootloader WIP * Reset log ID after fmt command * Added Read Out Protection support * Added CRC support class WIP * Added Log printouts

+++ Modules/Inc/BuzzerAgent.h
@@ -14,7 +14,7 @@
 typedef struct sBuzzer_Pattern
 {
     uint8_t patternLength;
-    uint8_t patternOnOff[16];
+    uint16_t patternOnOff[16];
     uint16_t patternDuration[16];
     bool isSingleCycle;
     uint8_t id;
